[Feat] 패링 구현#2
- 패링 실패시 데미지를 입도록 구현했습니다.
- parryDistanceOffset은 패링가능 거리에 실패로 처리하기위해 추가로 탐지하는 거리로, 패링 가능거리 오브젝트가 0이지만 추가 거리에서 탐지된 오브젝트가 하나이상이면 데미지를 입도록 구현했습니다.
- 데미지 처리 부분을 함수로 분리했습니다.

+++ Assets/_Scripts/Projectiles/Bullet.cs
@@ -6,15 +6,15 @@ public class Bullet : MonoBehaviour, IParryable
     [SerializeField] private float speed = 1f;
     [SerializeField] private float lifetime = 3f;
 
-    [SerializeField] private float damage = 1f;
+    [SerializeField] private int damage = 1;
     [SerializeField] private List<GameObject> ignoreObjects = new List<GameObject>(); //충돌 처리 무시할 오브젝트 등록(본인, 무기등)
 
-    public float Damage => damage;
+    public int Damage => damage;
 
 
     public bool CanBeParried { get; private set; } = true;
 
-    public void Init(float dmg)
+    public void Init(int dmg)
     {
         damage = dmg;
     }
